fix issues

+++ LRU.h
@@ -29,7 +29,7 @@ public:
         Node* tail;
         uint32_t salt1;
         uint32_t salt2;
-        uint32_t checkMagic;
+        uint64_t checkMagic;
 
 
         void insertStart(Node* newNode){
